# db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"database/sql"
	"time"
)

type OperationLog struct {
	ID          int64         `json:"id"`
	UserID      sql.NullInt64 `json:"user_id"`
	Username    string        `json:"username"`
	Method      string        `json:"method"`
	Path        string        `json:"path"`
	RequestBody string        `json:"request_body"`
	StatusCode  int32         `json:"status_code"`
	RequestID   string        `json:"request_id"`
	CreatedAt   time.Time     `json:"created_at"`
}

type User struct {
	ID             int64     `json:"id"`
	Username       string    `json:"username"`
	Email          string    `json:"email"`
	CreatedAt      time.Time `json:"created_at"`
	HashedPassword string    `json:"hashed_password"`
}

type Wallet struct {
	ID        int64     `json:"id"`
	UserID    int64     `json:"user_id"`
	Balance   int64     `json:"balance"`
	CreatedAt time.Time `json:"created_at"`
}