# docs/swagger.yaml
basePath: /
definitions:
  api.AuthUser:
    properties:
      email:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  api.PageResult:
    properties:
      list: {}
      pageNum:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  api.Response:
    properties:
      code:
        $ref: '#/definitions/errcode.Code'
      data: {}
      msg:
        type: string
    type: object
  api.createUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  api.createUserResponse:
    properties:
      user:
        $ref: '#/definitions/api.userResponse'
      wallet:
        $ref: '#/definitions/db.Wallet'
    type: object
  api.loginRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.loginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/api.userResponse'
    type: object
  api.operationLogResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      method:
        type: string
      path:
        type: string
      request_body:
        type: string
      request_id:
        type: string
      status_code:
        type: integer
      user_id:
        type: integer
      username:
        type: string
    type: object
  api.userResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  db.Wallet:
    properties:
      balance:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      user_id:
        type: integer
    type: object
  errcode.Code:
    enum:
    - 0
    - 10001
    - 10002
    - 10003
    - 10004
    - 10005
    - 20001
    - 20002
    - 20003
    - 20004
    - 30001
    type: integer
    x-enum-varnames:
    - Success
    - ErrInternal
    - ErrInvalidParams
    - ErrUnauthorized
    - ErrNotFound
    - ErrTimeout
    - ErrUserNotFound
    - ErrUserExists
    - ErrWrongCredentials
    - ErrTooManyLoginFails
    - ErrWalletNotFound
host: localhost:8080
info:
  contact: {}
  description: Golang Web 專案模板 API 文件。所有回應皆為統一格式 {code, msg, data}。
  title: template_golang_web API
  version: "1.0"
paths:
  /healthz:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 健康檢查
      tags:
      - system
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 帳號密碼
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.loginResponse'
              type: object
      summary: 登入
      tags:
      - auth
  /logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - TokenAuth: []
      summary: 登出
      tags:
      - auth
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.AuthUser'
              type: object
      security:
      - TokenAuth: []
      summary: 取得登入者資訊
      tags:
      - auth
  /operation-logs:
    get:
      parameters:
      - description: 頁碼（從 1 開始）
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 每頁筆數（5-50）
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/api.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/api.operationLogResponse'
                        type: array
                    type: object
              type: object
      security:
      - TokenAuth: []
      summary: 操作日誌列表
      tags:
      - system
  /users:
    get:
      parameters:
      - description: 頁碼（從 1 開始）
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 每頁筆數（5-50）
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/api.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/api.userResponse'
                        type: array
                    type: object
              type: object
      security:
      - TokenAuth: []
      summary: 使用者列表
      tags:
      - user
    post:
      consumes:
      - application/json
      parameters:
      - description: 使用者資料
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.createUserResponse'
              type: object
      summary: 建立使用者
      tags:
      - user
  /users/{id}:
    get:
      parameters:
      - description: 使用者 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.userResponse'
              type: object
      security:
      - TokenAuth: []
      summary: 查詢使用者
      tags:
      - user
  /wallet:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/db.Wallet'
              type: object
      security:
      - TokenAuth: []
      summary: 查詢自己的錢包
      tags:
      - wallet
securityDefinitions:
  TokenAuth:
    description: 登入後取得的 token，需驗證的 API 都要在 header 帶上
    in: header
    name: token
    type: apiKey
swagger: "2.0"